cleanups, games work with ISL, now we just need to clean up better when the ISL connection dies

+++ servatrice/src/servatrice.h
@@ -127,8 +127,8 @@ public:
 
 	QList<ServerProperties> getServerList() const;
 protected:
-	int startSession(const QString &userName, const QString &address);
-	void endSession(int sessionId);
+	qint64 startSession(const QString &userName, const QString &address);
+	void endSession(qint64 sessionId);
 	bool userExists(const QString &user);
 	AuthenticationResult checkUserPassword(Server_ProtocolHandler *handler, const QString &user, const QString &password, QString &reasonStr);
